[integ-tests] Add more deep learning AMIs in "remarkable" AMI list of build-image test
build-image test only use AL2 remarkable AMI. This commit add a '*' in the name to make sure ARM AMI can be retrieved too. We discovered the name was not enough for ARM from this PR https://github.com/aws/aws-parallelcluster/pull/6924

Moreover, this commit completes the list with deep learning AMIs for AL2023 and Ubuntu. Although they are not used in the tests, having them in the code is convenient for the future.

The Deep Learning AMI names are from https://docs.aws.amazon.com/dlami/latest/devguide/appendix-ami-release-notes-base.html